# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily sales report
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly sales report
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get detailed sales report with filters
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
        - name: paymentType
          in: query
          schema:
            type: string
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
...
  /reports/users:
    get:
      operationId: getUserPerformanceReport
      tags: [reports]
      summary: Get user performance report
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User performance report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPerformance"
  /reports/profit:
    get:
      operationId: getProfitReport
      tags: [reports]
      summary: Get profit report
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitReport"
  /reports/losses:
    get:
      operationId: getLossReport
      tags: [reports]
      summary: Get inventory loss report (returns, damaged, lost)
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Loss report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryLoss"

  # Documents
...
components:
  schemas:
...
    SaleItem:
      type: object
      required: [id, saleId, productId, quantity, unitPrice, subtotal]
      properties:
        id:
          type: integer
        saleId:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
    Sale:
      type: object
      required: [id, totalAmount, discount, amountPaid, change, paymentType, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        cashierName:
          type: ["string", "null"]
        totalAmount:
          type: number
        discount:
          type: number
        amountPaid:
          type: number
        change:
          type: number
        paymentType:
          type: string
          enum: [cash, wallet, bank]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
...
    DailyReport:
      type: object
      required: [date, totalSales, totalRevenue, totalProfit, totalTransactions, salesByCategory]
      properties:
        date:
          type: string
        totalSales:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalTransactions:
          type: integer
        salesByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySales"
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/ProductSales"
    MonthlyReport:
      type: object
      required: [year, month, totalRevenue, totalProfit, totalTransactions, dailyBreakdown]
      properties:
        year:
          type: integer
        month:
          type: integer
        totalRevenue:
          type: number
        totalProfit:
          type: number
        totalTransactions:
          type: integer
        dailyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DailySummary"
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/ProductSales"
    SalesReport:
      type: object
      required: [totalTransactions, totalRevenue, totalDiscount, totalProfit, salesByPaymentType, sales]
      properties:
        totalTransactions:
          type: integer
        totalRevenue:
          type: number
        totalDiscount:
          type: number
        totalProfit:
          type: number
        salesByPaymentType:
          type: array
          items:
            $ref: "#/components/schemas/PaymentTypeSummary"
        sales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
    PaymentTypeSummary:
      type: object
      required: [paymentType, count, total]
      properties:
        paymentType:
          type: string
        count:
          type: integer
        total:
          type: number
...
    UserPerformance:
      type: object
      required: [userId, userName, totalSales, totalRevenue, totalTransactions]
      properties:
        userId:
          type: integer
        userName:
          type: string
        totalSales:
          type: number
        totalRevenue:
          type: number
        totalTransactions:
          type: integer
    ProfitReport:
      type: object
      required: [totalRevenue, totalCost, totalProfit, profitMargin, byProduct]
      properties:
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalProfit:
          type: number
        profitMargin:
          type: number
        byProduct:
          type: array
          items:
            $ref: "#/components/schemas/ProductProfit"
    ProductProfit:
      type: object
      required: [productId, productName, quantitySold, revenue, cost, profit]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantitySold:
          type: integer
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
    CategorySales:
      type: object
      required: [category, revenue, count]
      properties:
        category:
          type: string
        revenue:
          type: number
        count:
          type: integer
    ProductSales:
      type: object
      required: [productId, productName, quantity, revenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        revenue:
          type: number
    DailySummary:
      type: object
      required: [date, revenue, transactions]
      properties:
        date:
          type: string
        revenue:
          type: number
        transactions:
          type: integer

    # Documents
...
    InventoryLoss:
      type: object
      required: [id, type, productId, quantity, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [return, damaged, lost]
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        reason:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
...