# lib/api-zod/src/generated/types/paymentTypeSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentTypeSummary {
  paymentType: string;
  count: number;
  total: number;
}