# lib/api-zod/src/generated/types/userPerformance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserPerformance {
  userId: number;
  userName: string;
  totalSales: number;
  totalRevenue: number;
  totalTransactions: number;
}